Moved mnethods to service to handle the form search
