+++ src/app/search/search-filter/search-filter.component.ts
@@ -2,6 +2,7 @@ import * as _ from 'lodash';
 import { Component, OnInit, ViewChild, Output, EventEmitter } from '@angular/core';
 import {TranslateService} from '@ngx-translate/core';
 import { NgForm } from '@angular/forms';
+import {SearchService} from '../search.service';
 
 @Component({
   selector: 'app-search-filter',
@@ -10,10 +11,10 @@ import { NgForm } from '@angular/forms';
 })
 export class SearchFilterComponent implements OnInit {
     @ViewChild('f', { static: false }) searchForm: NgForm;
-    @Output() resultObj:any = new EventEmitter<any>();
     journeyDetails:Array<any> = [];
     passengerNumber = 1;
-    constructor(private translate: TranslateService) {}
+    constructor(private translate: TranslateService,
+        private search: SearchService) {}
     
     ngOnInit() {}
     getSearchResult() {
@@ -21,15 +22,15 @@ export class SearchFilterComponent implements OnInit {
         let destination = (this.searchForm.value.destination) ? this.searchForm.value.destination : 'f25dvk';
         let traveldate = (!_.isEmpty(this.searchForm.value.travelDate)) ? this.formatDate(this.searchForm.value.travelDate) : this.getCurrentDate();
         let passenger = (this.searchForm.value.passenger) ? this.searchForm.value.passenger : '1';
-        this.resultObj.emit({origin: source, 
-                            destination: destination,
-                            traveldate: traveldate,
-                            passenger: passenger});
+        this.search.updateSearchResult({origin: source, 
+                                    destination: destination,
+                                    traveldate: traveldate,
+                                    passenger: passenger})
     }
 
     private getCurrentDate() {
         let today = new Date();
-        let day = (today.getMinutes() < 10) ? `0${today.getMinutes()}` : today.getMinutes();
+        let day = (today.getDate() < 10) ? `0${today.getDate()}` : today.getDate();
         return `${today.getFullYear()}-${(today.getMonth()+1)}-${day}`;
     }
