+++ src/app/search/search-result/search-result.component.less
@@ -175,3 +175,11 @@
     color: #6c757d
 }
 
+.no-result {
+    color: #502a09;
+    display: flex;
+    text-align: center;
+    justify-content: center;
+    padding-top: 20px;
+    width: 100%;
+}
