+++ src/app/search/search.component.less
@@ -17,10 +17,3 @@
     display: flex;
     flex-wrap: wrap;
 }
-.no-result {
-    color: #502a09;
-    display: flex;
-    text-align: center;
-    justify-content: center;
-    padding-top: 20px;
-}
